created FeeMewWrapper component and use in MewsFeed
