+++ ui/src/pages/MewsFeed.vue
@@ -54,54 +54,23 @@
         v-for="(mew, index) in mewsFeed"
         :key="index"
       >
-        <q-item-section avatar>
-          <agent-avatar
-            :agent-pub-key="authorPubKey(mew.feedMew.header.author)"
-            size="50"
-            class="cursor-pointer"
-            @click="onAgentClick(authorPubKey(mew.feedMew.header.author))"
-            @mouseenter="showProfile(index)"
-            @mouseleave="hideProfile(index)"
-          >
-            <q-menu
-              v-model="profileVisible[index]"
-              anchor="bottom middle"
-              self="top middle"
-              :offset="[-10, 5]"
-              no-focus
-            >
-              <ProfilePopup
-                :agent-pub-key="authorPubKey(mew.feedMew.header.author)"
-                @mouseenter="keepShowingProfile(index)"
-                @mouseleave="hideProfile(index)"
-              />
-            </q-menu>
-          </agent-avatar>
-        </q-item-section>
-        <q-item-section>
-          <MewsFeedContent :mew-content="mew.feedMew" />
-        </q-item-section>
+        <FeedMewWrapper 
+          :mew="mew"
+          :index="index"
+        /> 
       </q-item>
     </q-list>
   </q-page>
 </template>
 
 <script setup lang="ts">
-import { HoloHashB64 } from '@holochain-open-dev/core-types';
 import { createMew, mewsFeed as getMewsFeed } from '../services/clutter-dna';
 import { onMounted, ref } from 'vue';
-import { FeedMewWithContext as FeedMew, Mew, CreateMewInput } from '../types/types';
+import { FeedMewWithContext as FeedMew, CreateMewInput } from '../types/types';
 import { showError } from '../utils/notification';
-import { authorPubKey } from "../utils/hash";
-import { useRouter } from 'vue-router';
 import AddMew from '../components/AddMew.vue';
-import MewsFeedContent from '../components/MewsFeedContent.vue';
-import ProfilePopup from "../components/ProfilePopup.vue";
+import FeedMewWrapper from '../components/FeedMewWrapper.vue';
 
-
-const PROFILE_SHOW_HIDE_DELAY = 400; // in ms
-
-const router = useRouter();
 
 const loading = ref(false);
 const mewsFeed = ref<FeedMew[]>([]);
@@ -128,29 +97,5 @@ onMounted(loadMewsFeed);
 const publishMew = async (newMew: CreateMewInput) => {
   await createMew(newMew);
   loadMewsFeed();
-};
-
-const onAgentClick = (agentPubKey: HoloHashB64) => {
-  router.push(`/profiles/${agentPubKey}`);
-};
-
-const showProfile = (index: number) => {
-  profileVisible.value = new Array(mewsFeed.value.length).fill(false);
-  profileShowTimeouts.value[index] = window.setTimeout(
-    () => profileVisible.value[index] = true,
-    PROFILE_SHOW_HIDE_DELAY
-  );
-};
-
-const hideProfile = (index: number) => {
-  window.clearTimeout(profileShowTimeouts.value[index]);
-  profileHideTimeouts.value[index] = window.setTimeout(
-    () => profileVisible.value[index] = false,
-    PROFILE_SHOW_HIDE_DELAY
-  );
-};
-
-const keepShowingProfile = (index: number) => {
-  window.clearTimeout(profileHideTimeouts.value[index]);
 };
 </script>
